+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [1.0.1]
+- Add file
+  - Add logo for vsCode extension
+  - add preview command for twig file
+
 ## [1.0.0]
 - Add new features !
   - set HTML emet by default in twig file